Move discontinued cards to archive
I just got off the phone with BoA trying to do a PC. It seems like many of the previously existing cards were discontinued recently and the website has not been updated to reflect that. When the rep tries to actually do the product change, the card no longer appears as an option.